+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## [4.0.0] - 2017-11-07
+
+### Changed
+
+* Updated `sendSms`, `sendEmail` and `sendLetter` to take an `options` object as a parameter.
+    * `personalisation`, `reference`, `smsSenderId` and `emailReplyToId` now need to be passed to these functions inside `options`.
+
 ## [3.5.0] - 2017-11-01
 
 ### Changed